split internalNames constants to new file
