+++ src/client/src/utils/constants/internalNames.ts
@@ -1,0 +1,32 @@
+const WIZARD_CONTENT_INTERNAL_NAMES = {
+  ANGULAR: "Angular",
+  APP_SERVICE: "wts.Feature.Azure.AppService",
+  AZURE: "wts.Feature.Azure",
+  REACT_BLANK_PAGE: "wts.Page.React.Blank",
+  REACT_CONTENT_GRID: "wts.Page.React.Grid",
+  REACT_MASTER_DETAIL: "wts.Page.React.MasterDetail",
+  REACT_LIST: "wts.Page.React.List",
+  ANGULAR_BLANK_PAGE: "wts.Page.Angular.Blank",
+  ANGULAR_CONTENT_GRID: "wts.Page.Angular.Grid",
+  ANGULAR_MASTER_DETAIL: "wts.Page.Angular.MasterDetail",
+  ANGULAR_LIST: "wts.Page.Angular.List",
+  COSMOS_DB: "wts.Feature.Azure.Cosmos",
+  COSMOS_DB_MONGO: "wts.Feature.Azure.Cosmos.Mongo",
+  COSMOS_DB_SQL: "wts.Feature.Azure.Cosmos.SQL",
+  FULL_STACK_APP: "FullStackWebApp",
+  NODE: "Node",
+  MOLECULER: "Moleculer",
+  ASPNET: "AspNet",
+  FLASK: "Flask",
+  REACT: "React",
+  REST_API: "RestAPI",
+  VUE: "Vue",
+  VUE_BLANK_PAGE: "wts.Page.Vue.Blank",
+  VUE_CONTENT_GRID: "wts.Page.Vue.Grid",
+  VUE_MASTER_DETAIL: "wts.Page.Vue.MasterDetail",
+  VUE_LIST: "wts.Page.Vue.List"
+};
+
+export {
+  WIZARD_CONTENT_INTERNAL_NAMES,
+};
